refactor: repoint chart-config imports to @/ui and delete utils shim
