+++ src/components/ui/charts/AtlasBoxPlotChart.story.vue
@@ -4,7 +4,12 @@ const data = [{ category: 'Age', min: 0, p10: 5, p25: 20, median: 40, p75: 60, p
 </script>
 
 <template>
-  <Story title="Charts/AtlasBoxPlotChart" group="charts">
-    <Variant title="default"><Chart :data="data" /></Variant>
+  <Story
+    title="Charts/AtlasBoxPlotChart"
+    group="charts"
+  >
+    <Variant title="default">
+      <Chart :data="data" />
+    </Variant>
   </Story>
 </template>
