--- src/utils/chart-config.ts
@@ -1,3 +1,0 @@
-// Compatibility shim — chart-config moved to @/ui/chart-config.
-// Remove once all imports are migrated (see plan Task 14).
-export * from '@/ui/chart-config'
