+++ tests/unit/components/datasources/charts/DashboardGenderChart.spec.ts
@@ -5,7 +5,7 @@ import * as components from 'vuetify/components'
 import * as directives from 'vuetify/directives'
 import type { PieChartData } from '@/models/datasource.types'
 
-vi.mock('@/utils/chart-config', () => ({
+vi.mock('@/ui/chart-config', () => ({
   dashboardGenderPieOptions: vi.fn((data) => ({
     series: [{
       type: 'pie',
@@ -15,12 +15,12 @@ vi.mock('@/utils/chart-config', () => ({
   createResizeHandler: vi.fn(() => vi.fn())
 }))
 
-let chartConfig: typeof import('@/utils/chart-config')
+let chartConfig: typeof import('@/ui/chart-config')
 let DashboardGenderChart: typeof import('@/components/datasources/charts/DashboardGenderChart.vue').default
 
 beforeAll(async () => {
   vi.resetModules()
-  chartConfig = await import('@/utils/chart-config')
+  chartConfig = await import('@/ui/chart-config')
   DashboardGenderChart = (await import('@/components/datasources/charts/DashboardGenderChart.vue')).default
 })
