+++ tests/unit/components/reports/charts/TreemapChart.spec.ts
@@ -11,7 +11,7 @@ import TreemapChart from '@/components/ui/charts/AtlasTreemapChart.vue'
 import type { TreemapNode } from '@/models/report.types'
 
 // Mock chart config utilities
-vi.mock('@/utils/chart-config', () => ({
+vi.mock('@/ui/chart-config', () => ({
   defaultTreemapOptions: vi.fn((data, title) => ({
     title: { text: title },
     series: [{ type: 'treemap', data, roam: true }]
